A simple Password Locker GUI.
Made mainly with TKINTER and auto-py-to-exe. To enter the GUI you must first type the "MasterPassword" (default is 0000). All passwords must be encripted with 2 keys to avoid them easily accessed in the CSV (I'm aware of binary files, but I wanted to try something else). For some reason I was not able to make the pygame mixer work as an exe so there is no background music there :(
Also if you want to play the song anyway, you need to insert it as .wav.